Moved cleanString method to pb gen text
