+++ lib/generation/generators/visual-widgets/pb_text_gen.dart
@@ -1,4 +1,3 @@
-import 'package:parabeac_core/controllers/main_info.dart';
 import 'package:parabeac_core/generation/generators/import_generator.dart';
 import 'package:parabeac_core/generation/generators/pb_generator.dart';
 import 'package:parabeac_core/generation/generators/visual-widgets/pb_text_style_gen_mixin.dart';
@@ -17,6 +16,25 @@ class PBTextGen extends PBGenerator with PBTextStyleGen {
     'STRIKETHROUGH': 'TextDecoration.lineThrough',
   };
 
+  /// A cleaning map,
+  /// The value on the left will be replaced by the value on the right
+  /// They will be applied in order from top to bottom
+  static final Map<dynamic, String> _replaceMap = {
+    '\n': ' ',
+    '\'': '\\\'',
+    '\$': '\\\$',
+    RegExp(r'[\x00-\x1f]'): '',
+  };
+
+  /// Applies the replacement list to text
+  static String cleanString(String text) {
+    var newText = text;
+    _replaceMap.forEach((target, replacement) {
+      newText = newText.replaceAll(target, replacement);
+    });
+    return newText ?? '';
+  }
+
   static String getDecoration(String decoration) {
     if (_decorationMap.containsKey(decoration)) {
       return _decorationMap[decoration];
@@ -28,7 +46,7 @@ class PBTextGen extends PBGenerator with PBTextStyleGen {
   String generate(PBIntermediateNode source, PBContext context) {
     if (source is InheritedText) {
       var textStyle = source.auxiliaryData.intermediateTextStyle;
-      var cleanText = MainInfo.cleanString(source.text);
+      var cleanText = cleanString(source.text);
 
       context.project.genProjectData.addDependencies('auto_size_text', '3.0.0');
